% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tylervigen.R
\docType{data}
\name{colors}
\alias{colors}
\title{Internal Color Palette}
\format{An object of class \code{list} of length 3.}
\usage{
colors
}
\value{
list of named hex color codes
}
\description{
Internal color palette for the package.
}
\keyword{datasets}
