% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tylervigen.R
\name{print.ggplot_suppressed}
\alias{print.ggplot_suppressed}
\title{Explicitly draw plot}
\usage{
\method{print}{ggplot_suppressed}(x, ...)
}
\arguments{
\item{x}{object of class ggplot_suppressed}

\item{...}{other arguments not used by this method}
}
\value{
Invisibly returns the result of ggplot_build(), which is a list with components that contain the plot itself, the data, information about the scales, panels etc.
}
\description{
The ggplot2 code necessary to make plots in Tyler Vigen's style
require aggressive parameters for the splines which produce warning messages.
This custom print function suppresses these messages.
}
