% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tylervigen.R
\docType{data}
\name{tylervigen}
\alias{tylervigen}
\title{Spurious Data Collated by Tyler Vigen}
\format{A data frame with 15 rows and 31 variables:
\describe{
  \item{year}{year of given observation}

  \item{science_spending}{US spending on science, space, and technology in billions of dollars
  (\href{https://www.census.gov/compendia/statab/cats/science_technology/expenditures_research_development.html}{U.S. Office of Management and Budget})}

  \item{hanging_suicides}{suicides by hanging, strangulation, and suffocation
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{pool_fall_drownings}{number of people who drowned by falling into a pool
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{cage_films}{films Nicholas Cage appeared in
  (\href{https://www.imdb.com/name/nm0000115/}{Internet Movie Database})}

  \item{cheese_percap}{per capita cheese consumption
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0217.xls}{Department of Agriculture})}

  \item{bed_deaths}{number of people who died by becoming tangled in their bedsheets
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{maine_divorce_rate}{divorce rate in Maine
  (\href{https://www.census.gov/compendia/statab/cats/births_deaths_marriages_divorces/marriages_and_divorces.html}{National Vital Statistics Reports})}

  \item{margarine_percap}{per capita consumption of margarine
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0217.xls}{U.S. Department of Agriculture})}

  \item{miss_usa_age}{age of miss america
  (\href{https://en.wikipedia.org/wiki/Miss_America}{Wikipedia})}

  \item{steam_murders}{number of murders by steam, hot vapours, or hot objects
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control and Prevention})}

  \item{arcade_revenue}{total revenue generated by arcades in billions of dollars
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s1228.xls}{U.S. Census Bureau})}

  \item{computer_science_doctorates}{computer science doctorates awarded in the US
  (\href{https://www.nsf.gov/statistics/infbrief/nsf11305/}{National Science Foundation})}

  \item{noncom_space_launches}{worldwide non-commercial space launches
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0822.xls}{Federal Aviation Administration})}

  \item{sociology_doctorates}{sociology doctorates awarded (US)
  (\href{https://www.nsf.gov/statistics/infbrief/nsf11305/}{National Science Foundation})}

  \item{mozzarella_percap}{per capita consumtion of mozzarella cheese
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0217.xls}{U.S. Department of Agriculture})}

  \item{civil_engineering_doctorates}{number of civil engineering doctorates awarded
  (\href{https://www.nsf.gov/statistics/infbrief/nsf11305/}{National Science Foundation})}

  \item{fishing_drownings}{people who drowned after falling out of a fishing boat
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{kentucky_marriage_rate}{marriage rate in Kentucky
  (\href{https://www.census.gov/compendia/statab/cats/births_deaths_marriages_divorces/marriages_and_divorces.html}{National Vital Statistics Reports})}

  \item{train_collision_deaths}{drivers killed in collisions with railway train
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{oil_imports_norway}{US crude oil imports from Norway
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0937.pdf}{Department of Energy})}

  \item{chicken_percap}{per capita consumption of chicken
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0217.xls}{U.S. Department of Agriculture})}

  \item{oil_imports_total}{total US crude oil imports
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0937.pdf}{Department of Energy})}

  \item{pool_drownings}{number of people who drowned while in a swimming-pool
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{nuclear_power}{power generated by US nuclear power plants
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0937.pdf}{Department of Energy})}

  \item{japanese_cars_sold}{Japanese passenger cars sold in the US
  (\href{https://www.census.gov/compendia/statab/2008/tables/08s1026.xls}{U.S. Bureau of Transportation Statistics})}

  \item{motor_vehicle_suicides}{suicides by crashing of motor vehicle
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{spelling_bee_word_length}{letters in winning Word of Scripps National Spelling Bee
  (\href{https://www.spellingbee.com/champions-and-their-winning-words}{National Spelling Bee})}

  \item{spider_deaths}{number of people killed by venomous spiders
  (\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention})}

  \item{math_doctorates}{math doctorates awarded
  (\href{https://www.nsf.gov/statistics/infbrief/nsf11305/}{National Science Foundation})}

  \item{uranium}{uranium stored at US nuclear power plants
  (\href{https://www.census.gov/compendia/statab/2012/tables/12s0937.pdf}{Department of Engergy})}
}}
\source{
\url{https://www.tylervigen.com/spurious-correlations}
}
\usage{
tylervigen
}
\description{
A dataset containing information collated by Tyler Vigen from various
sources to showcase spurious relationships.
}
\keyword{datasets}
