% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tylervigen.R
\name{tylervigen_plot2}
\alias{tylervigen_plot2}
\title{Tyler Vigen Plot 2}
\usage{
tylervigen_plot2()
}
\description{
Number of people who drowned by falling into a pool correlates with Films Nicolas Cage appeared in (r = 0.67).
}
\references{
\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention}

  \href{https://www.imdb.com/name/nm0000115/}{Internet Movie Database}
}
\keyword{cage}
\keyword{drowning,}
\keyword{pool,}
