% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tylervigen.R
\name{tylervigen_plot3}
\alias{tylervigen_plot3}
\title{Tyler Vigen Plot 3}
\usage{
tylervigen_plot3()
}
\description{
Per capita cheese consumption correlates with Number of people who died by becoming tangled in their bedsheets (r = 0.95).
}
\references{
\href{https://www.census.gov/compendia/statab/2012/tables/12s0217.xls}{Department of Agriculture}

  \href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention}
}
\keyword{bed,}
\keyword{cheese,}
\keyword{death,}
\keyword{sheet}
