% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tylervigen.R
\name{tylervigen_plot9}
\alias{tylervigen_plot9}
\title{Tyler Vigen Plot 9}
\usage{
tylervigen_plot9()
}
\description{
People who drowned after falling out of a fishing boat correlates with marriage rate in Kentucky (r = 0.95).
}
\references{
\href{https://wonder.cdc.gov/}{Centers for Disease Control & Prevention}

  \href{https://www.census.gov/compendia/statab/cats/births_deaths_marriages_divorces/marriages_and_divorces.html}{National Vital Statistics Reports}
}
\keyword{Kentucky}
\keyword{boat,}
\keyword{drowning,}
\keyword{fishing,}
\keyword{marriage,}
