% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whysospurious-package.R
\docType{package}
\name{whysospurious-package}
\alias{whysospurious}
\alias{whysospurious-package}
\title{whysospurious}
\description{
A package for illustrating the golden rule that "correlation does not imply causation".
}
\details{
The whysospurious package provides raw data, visualization functions, and ggplot2 objects divided according to source.
}
\section{tylervigen objects}{

The tylervigen objects are sourced from \href{https://www.tylervigen.com/spurious-correlations}{Spurious Correlations}.
}

\author{
\strong{Maintainer}: Matthew Schuelke \email{matt@themadstatter.com} (\href{https://orcid.org/0000-0001-5755-1725}{ORCID})

Other contributors:
\itemize{
  \item Maya Gann-Bociek \email{maya.gannbociek@slu.edu} [contributor]
}

}
\keyword{internal}
